refacto(folios) refactor remove card from main folio
