--- tests/functional/controllers/card_folios_controller.spec.ts
@@ -1,115 +1,0 @@
-import { test } from '@japa/runner'
-import testUtils from '@adonisjs/core/services/test_utils'
-import sinon from 'sinon'
-import { FolioFactory } from '#database/factories/folio'
-import AuthServiceMock, { TEST_AUTH_USER_ID } from '#tests/mocks/auth_service_mock'
-import { CardFactory } from '#database/factories/card'
-import CardFolio from '#models/card_folio'
-import { CardFolioFactory } from '#database/factories/card_folio'
-import { ArtistFactory } from '#database/factories/artist'
-import { RarityFactory } from '#database/factories/rarity'
-import { LegalityFactory } from '#database/factories/legality'
-import { SetFactory } from '#database/factories/set'
-
-test.group('Card Folio Controller', (group) => {
-  let wardenApiClientStub: sinon.SinonStub
-
-  group.setup(() => {
-    wardenApiClientStub = AuthServiceMock.setupWardenApiClientStub()
-  })
-
-  group.each.setup(() => testUtils.db().withGlobalTransaction())
-
-  group.teardown(() => {
-    wardenApiClientStub.restore()
-  })
-
-  test('delete - it should remove a card from user main folio', async ({ client, assert }) => {
-    const userId = TEST_AUTH_USER_ID
-
-    const rootFolio = await FolioFactory.merge({
-      userId,
-      name: 'root',
-      isRoot: true,
-    }).create()
-
-    await ArtistFactory.create()
-    await RarityFactory.create()
-    await LegalityFactory.create()
-    await SetFactory.merge({ id: 'base1' }).create()
-    const card = await CardFactory.create()
-
-    await CardFolioFactory.merge({
-      cardId: card.id,
-      folioId: rootFolio.id,
-      occurrence: 2,
-    }).create()
-
-    const response = await client
-      .delete(`/api/v1/folios/cards/${card.id}`)
-      .header('Authorization', 'Bearer fake-token-for-testing')
-
-    response.assertStatus(200)
-    response.assertBodyContains({
-      message: 'Card remove from main folio successfully',
-    })
-
-    const deletedCardFolio = await CardFolio.query()
-      .where('folio_id', rootFolio.id)
-      .where('card_id', card.id)
-      .first()
-
-    assert.isNull(deletedCardFolio)
-  })
-
-  test('delete - it should return 404 when card folio does not exist', async ({ client }) => {
-    const userId = TEST_AUTH_USER_ID
-
-    await FolioFactory.merge({
-      userId,
-      name: 'root',
-      isRoot: true,
-    }).create()
-
-    await ArtistFactory.create()
-    await RarityFactory.create()
-    await LegalityFactory.create()
-    await SetFactory.merge({ id: 'base1' }).create()
-    const card = await CardFactory.create()
-
-    const response = await client
-      .delete(`/api/v1/folios/cards/${card.id}`)
-      .header('Authorization', 'Bearer fake-token-for-testing')
-
-    response.assertStatus(404)
-    response.assertBodyContains({
-      code: 'E_ROW_NOT_FOUND',
-    })
-  })
-
-  test('delete - it should return 404 for non-existent card', async ({ client }) => {
-    const userId = TEST_AUTH_USER_ID
-
-    await FolioFactory.merge({
-      userId,
-      name: 'root',
-      isRoot: true,
-    }).create()
-
-    const response = await client
-      .delete('/api/v1/folios/cards/non-existente')
-      .header('Authorization', 'Bearer fake-token-for-testing')
-
-    response.assertStatus(404)
-    response.assertBodyContains({
-      message: 'Card not found',
-      code: 'E_ROW_NOT_FOUND',
-    })
-  })
-
-  test('delete - it should require authentication', async ({ client }) => {
-    const response = await client.delete('/api/v1/folios/cards/base1-1')
-
-    response.assertStatus(401)
-  })
-})
